amd64/requirements.txt
numpy==1.21.4
pandas==1.3.4
joblib==1.1.0
matplotlib==3.5.0 
sklearn==0.0
tqdm==4.62.3
kafka-python==2.0.2
google-cloud==0.34.0
google-auth==2.3.3
google-cloud-storage==1.42.3